LayoutTests/platform/chromium/plugins/gesture-events-scrolled-expected.txt

Plugin received event: GestureTap
Plugin received event: MouseEnter
Plugin received event: MouseMove
Plugin received event: MouseDown
Plugin received event: MouseUp
Plugin received event: GestureScrollBegin
Plugin received event: GestureScrollUpdate
Plugin received event: GestureScrollUpdate
Plugin received event: GestureScrollEnd
